Modified implementation of DtcastPackets Basic implementations of Source Routing, Forwarding, MessageQueue tables Beginning to implement actual click Elements

+++ elements/dtcast/Forwarder.cc
@@ -10,12 +10,19 @@
 CLICK_DECLS
 
 DtcastForwarder::DtcastForwarder()
+		: _source(NULL), _receiver(NULL)
+{
+}
+
+DtcastForwarder::~DtcastForwarder( )
 {
 }
 
 int DtcastForwarder::configure( Vector<String> &conf, ErrorHandler *errH )
 {
 	return cp_va_kparse( conf,this,errH,
+			"DTCAST_SOURCE",   cpkPositional, cpElement, &_source,
+			"DTCAST_RECEIVER", cpkPositional, cpElement, &_receiver,
 //					"TCP_DROP",  cpkPositional, cpInteger, &_tcpDrop,
 //					"TCP_RESET", cpkPositional, cpInteger, &_tcpReset,
 //					"TCP_LONG_RESET",  cpkPositional, cpInteger, &_tcpLongReset,
@@ -30,12 +37,17 @@ int DtcastForwarder::configure( Vector<String> &conf, ErrorHandler *errH )
 Packet* DtcastForwarder::simple_action( Packet *pkt )
 {
 	//pkt should containt valid IP packet with protocol field set to IP_PROTO_DTCAST
-	pkt->kill( );
+	DtcastPacket *dtcast=DtcastPacket::make( pkt );
+	if( dtcast==NULL )
+	{
+		pkt->kill( );
+		return NULL;
+	}
 	
 	return NULL;
 }
 
-		
+
 CLICK_ENDDECLS
 EXPORT_ELEMENT(DtcastForwarder)
